gralloc: Add and use new flag for exynos7580-specific code
The previous conditions guarding the code were not right at all.

I somehow forgot to check if they were fine for devices with other SoCs,
and as it turns out, the "if TARGET_SOC == exynos7580" conditions, for
example, were somehow always true even if TARGET_SOC was not exynos7580.
That resulted in the 7580-specific code being incorrectly used.

So, as a proper solution, add a new flag, USES_EXYNOS7580, that is set
if TARGET_SOC is actually exynos7580, and replace the conditions with
ones checking it.

Change-Id: Id7d75a6265bdb54bf9db0b192f5c254a0ced8cfb